requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
jinja2==3.1.2
sqlmodel==0.0.8
python-multipart==0.0.6
python-dateutil==2.8.2
aiofiles==23.1.0
pytest==7.4.0
